requirements.txt
asgiref==3.2.3
astroid==2.3.3
coverage==5.0.3
Django==3.0.3
django-autofixture==0.12.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-jenkins==0.110.0
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
drf-generators==0.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.4
pyflakes==2.2.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
